+++ rust/crates/entrypoint/src/interactive/mod.rs
@@ -1,0 +1,33 @@
+mod apply_processed;
+mod build_group_imports;
+mod build_imports_from_existing;
+mod build_policy_imports;
+mod build_resource_group_imports;
+mod build_role_assignment_imports;
+mod clean_all;
+mod clean_imports;
+mod clean_processed;
+mod init_processed;
+mod jump_to_block;
+mod list_imports;
+mod open_dir;
+mod pim_activate;
+mod populate_cache;
+
+pub mod prelude {
+    pub use crate::interactive::apply_processed::*;
+    pub use crate::interactive::build_group_imports::*;
+    pub use crate::interactive::build_imports_from_existing::*;
+    pub use crate::interactive::build_policy_imports::*;
+    pub use crate::interactive::build_resource_group_imports::*;
+    pub use crate::interactive::build_role_assignment_imports::*;
+    pub use crate::interactive::clean_all::*;
+    pub use crate::interactive::clean_imports::*;
+    pub use crate::interactive::clean_processed::*;
+    pub use crate::interactive::init_processed::*;
+    pub use crate::interactive::jump_to_block::*;
+    pub use crate::interactive::list_imports::*;
+    pub use crate::interactive::open_dir::*;
+    pub use crate::interactive::pim_activate::*;
+    pub use crate::interactive::populate_cache::*;
+}