+++ rust/crates/entrypoint/src/lib.rs
@@ -1,8 +1,9 @@
 #![feature(let_chains, async_closure, iter_collect_into, duration_constructors)]
-mod action;
-mod actions;
 mod clap;
+mod interactive;
+mod menu_action;
 mod menu;
+mod noninteractive;
 mod read_line;
 mod version;
 pub mod prelude {