+++ rust/crates/entrypoint/src/noninteractive/mod.rs
@@ -1,0 +1,8 @@
+mod resource_group_imports;
+mod perform_import;
+mod process_generated;
+pub mod prelude {
+    pub use crate::noninteractive::resource_group_imports::*;
+    pub use crate::noninteractive::perform_import::*;
+    pub use crate::noninteractive::process_generated::*;
+}