+++ rust/changelog.md
@@ -3,6 +3,7 @@
 - Fix terminal colours in default terminal opened when double clicking the exe
     - https://stackoverflow.com/questions/78741673/colors-not-working-on-default-terminal-for-release-rust-exe/78741674#78741674
 - Add app icon
+- Clean up non-interactive usage scenarios (see: `cloud_terrastodon --help`)
 
 # v0.1.1